GTK3: Everything is deprecated!
Port GtkTable to GtkGrid.
Port from GtkStock to translatable strings.
Ignore the GtkUIManager/GtkAction stuff for now. It's a lot of
work and there's no gain. We'll need to switch to GtkApplication
and use GIO's GMenu in the future. But we don't need to export
the menu over DBUS or anything silly like that so we don't need
to port it now.